requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.1
python-dotenv==1.0.0 
#列出了项目所需的所有依赖包及其版本